+++ internal/models/product.go
