+++ internal/services/user.go
@@ -3,7 +3,7 @@ package services
 import (
 	"context"
 
-	"github.com/josuetorr/frequent-flyer/server/models"
+	"github.com/josuetorr/frequent-flyer/internal/models"
 )
 
 type UserService struct {
